medtls: add interfaces to support new interface of optee os
optee os add new interface for context init and free. With these
interfaces, we do not need to add supporting external memory allocate
for mbedTLS. But, it is still needed some changing for mbedtls.
1. add mbedtls_cipher_clone() for cipher to copy context between two
operations.
2. add mbedtls_cipher_setup_info() for cipher. cipher need know the key
length to find its 'cipher_info', but it is not input in allocate
function. So, use a default key len in the begin. It need to reset the
cipher info again in init function.
3. add mbedtls_cipher_cmac_setup() for cmac. this function is separate
from mbedtls_cipher_cmac_starts().
4. copy hmac context in md.

Signed-off-by: Edison Ai <edison.ai@arm.com>